add jwt add

+++ index.js
@@ -5,6 +5,7 @@ require('dotenv').config()
 const app = express()
 const port = process.env.PORT || 5000;
 
+
 //meadile wear
 
 app.use(cors())
@@ -33,7 +34,7 @@ function jwtVerify(req, res, next) {
 
     const authorisation = req.headers.authorisation
     if (!authorisation) {
-        return res.status(401).body({
+        return res.status(401).send({
             success: false,
             message: 'unauthorised access 401'
         })
@@ -48,8 +49,8 @@ function jwtVerify(req, res, next) {
             })
 
         }
-       req.decord=decord
-       next()
+        req.decord = decord
+        next()
     })
 
 }
@@ -61,6 +62,7 @@ function jwtVerify(req, res, next) {
 app.post('/jwt', (req, res) => {
     const user = req.body;
     const token = jwt.sign(user, process.env.JWT_TOKEN_KEY, { expiresIn: '20h' })
+    
     res.send({ token })
 })
 
@@ -69,7 +71,7 @@ app.post('/jwt', (req, res) => {
 app.get('/delivery', async (req, res) => {
     try {
         if (req?.headers?.pages == 3) {
-            const result = await serviceCollaction.find({}).sort({_id:-1}).limit(3).toArray()
+            const result = await serviceCollaction.find({}).sort({ _id: -1 }).limit(3).toArray()
 
             res.send({
                 success: true,
@@ -127,7 +129,7 @@ app.post('/delivery', async (req, res) => {
 
 
     try {
-       
+
 
         const result = await serviceCollaction.insertOne(data)
         // console.log(result)
@@ -144,7 +146,7 @@ app.post('/delivery', async (req, res) => {
                 message: `Do not input Data`
             })
         }
-        console.log(result);
+        
 
 
     } catch (error) {
@@ -160,12 +162,18 @@ app.post('/delivery', async (req, res) => {
 
 // Add customer reviews 
 
-app.post('/reviews', async (req, res) => {
+app.post('/reviews', jwtVerify, async (req, res) => {
     const data = req.body
     // console.log(data)
     // const data = {email : 'sampodnath'}
-
+  
     try {
+        if (req.headers.email !== req.decord.email) {
+            return res.status(401).send({
+                success: false,
+                message: 'you are not assess this data'
+            })
+        }
         const result = await reviewCollaction.insertOne(data)
 
         if (result.insertedId) {
@@ -180,7 +188,7 @@ app.post('/reviews', async (req, res) => {
                 message: `Do not input Data`
             })
         }
-        console.log(result);
+      
 
 
     } catch (error) {
@@ -192,10 +200,18 @@ app.post('/reviews', async (req, res) => {
     }
 })
 
-app.get('/reviews/:email', async (req, res) => {
-const email = req.params.email
+app.get('/reviews/:email', jwtVerify, async (req, res) => {
+    const email = req.params.email
+
     try {
-        const result = await reviewCollaction.find({email}).sort({time:-1}).toArray()
+        if (email !== req.decord.email) {
+            return res.status(401).send({
+                success: false,
+                message: 'you are not assess this data'
+            })
+
+        }
+        const result = await reviewCollaction.find({ email }).sort({ time: -1 }).toArray()
         // console.log(result)
         res.send({
             success: true,
@@ -212,11 +228,11 @@ const email = req.params.email
 })
 
 app.get('/reviewsId/:id', async (req, res) => {
-const productId = req.params.id
-console.log(productId);
+    const productId = req.params.id
+
     try {
-        const result = await reviewCollaction.find({productId}).toArray()
-        console.log(result)
+        const result = await reviewCollaction.find({ productId }).toArray()
+   
         res.send({
             success: true,
             message: 'successfull gate data',
@@ -231,9 +247,15 @@ console.log(productId);
     }
 })
 
-app.delete('/reviews/:id', async (req, res) => {
+app.delete('/reviews/:id', jwtVerify, async (req, res) => {
     const quaryid = req.params.id
     try {
+        if (req.headers.email !== req.decord.email) {
+            return res.status(401).send({
+                success: false,
+                message: 'you are not assess this data'
+            })
+        }
 
         const result = await reviewCollaction.deleteOne({ _id: ObjectId(quaryid) })
         if (result.deletedCount === 1) {
@@ -259,11 +281,17 @@ app.delete('/reviews/:id', async (req, res) => {
     }
 })
 
-app.patch('/reviews/:id', async (req, res) => {
+app.patch('/reviews/:id', jwtVerify, async (req, res) => {
     const { id } = req.params
     try {
+        if (req.headers.email !== req.decord.email) {
+            return res.status(401).send({
+                success: false,
+                message: 'you are not assess this data'
+            })
+        }
         const result = await reviewCollaction.updateOne({ _id: ObjectId(id) }, { $set: req.body })
-        console.log(result)
+
         if (result.matchedCount) {
             res.send({
                 success: true,
@@ -285,8 +313,9 @@ app.patch('/reviews/:id', async (req, res) => {
     }
 })
 
-app.get('/reviews/:id', async (req, res) => {
+app.get('/reviewsSongle/:id', async (req, res) => {
     const id = req.params.id
+    console.log(id);
     try {
         const result = await reviewCollaction.findOne({ _id: ObjectId(id) })
